+++ cpu.h
@@ -32,7 +32,7 @@ typedef struct {
 } Registers;
 
 int execute(FILE *program, uint8_t *ram);
-int readByte(FILE* program, uint8_t pc);
+uint8_t readByte(FILE* program, uint8_t pc);
 void step(uint8_t opcode, FILE* program, uint8_t *ram, Registers *registers);
 void notImplemented(uint8_t opcode);
 void illegalOpcode(uint8_t opcode);
